// lifetrace-frontend/lib/generated/schemas/noteLinkResponse.ts
/**
 * Generated by orval v8.2.0 🍺
 * Do not edit manually.
 * LifeTrace API
 * LifeTrace API
 * OpenAPI spec version: 0.1.2
 */
import type { NoteCounterpart } from './noteCounterpart';
import type { NoteLinkResponseRelationType } from './noteLinkResponseRelationType';

/**
 * 思想链接响应
 */
export interface NoteLinkResponse {
  id: number;
  source_note_id: number;
  target_note_id: number;
  relation_type: NoteLinkResponseRelationType;
  user_note?: string | null;
  created_at: string;
  counterpart?: NoteCounterpart | null;
}