// lifetrace-frontend/lib/generated/schemas/noteLinkResponseRelationType.ts
/**
 * Generated by orval v8.2.0 🍺
 * Do not edit manually.
 * LifeTrace API
 * LifeTrace API
 * OpenAPI spec version: 0.1.2
 */

export type NoteLinkResponseRelationType = typeof NoteLinkResponseRelationType[keyof typeof NoteLinkResponseRelationType];


export const NoteLinkResponseRelationType = {
  SUPPORTS: 'SUPPORTS',
  EXTENDS: 'EXTENDS',
  CONTRADICTS: 'CONTRADICTS',
  RELATES: 'RELATES',
} as const;